backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.2
openai==1.14.0
plotly==5.19.0
pulp==2.7.0
networkx>=2.8.0
python-multipart==0.0.9
aiofiles==23.2.1 